WHISP_TPW-PROJECT2/Whisp/requirements.txt
Django

djangorestframework
markdown
django-filter
django-cors-headers

django==4.2.5
django-allauth
sorl-thumbnail
pillow
django-notifications-hq
dj_rest_auth

tslib